Refactor to v2
With all indications going towards the logic of the whole system pushing
us towards lots and lots of components, the old storage method of
allocating a component for each entity, whether it is allocated or not
seems too much.

So we're changing the data organization to a two-stage table system.

Seperates data storage into an entity table and many component tables
that store one or more components.

Each Entity row has two parts, each with as many values as there are
tables. The first half contains the component masks and the second the
row offsets.

The row offset gives the offset of the component row within its table.

Names the entity table "e" and the component tables c0 to c9 providing
one-step array acquisition from the entity System.

In general things get a bit more complicated for the macro. Since we now
have to look up the row offset for each entity, we need to make sure
that we keep that offset up-to-date. The macro now updates the
corresponding variable after a js assignment statement to the variable

Statements like (entity+1).x are not supposed and not caught by the
macro member access logic.

This seems preferable to doing the lookup for each entity prop access.