Implement JSON exporting and importing
Adds a system.toJSON([masks]) method to create a JSON object graph from
the current entity system, optionally filtered by an array of mask
values per table.

Creates EntitySystem.fromJSON(config, json) to create a new entity
system from such an JSON object graph export.

In general, the entity objects contained in the "entities" array of the
export are exactly in the format they need to be to be used as template.

For this purpose we extend our template logic by two fields:

A field "_id" may contain the id of an entity. system.addComponents()
will however just ignore that field. If you want to recreate an entity,
read the _id field from the export and call
entitySysten.newEntity(template, _id).

A field "_" is generated by the export and contains the name of every
tag component the entity has. Giving a template with such a "_" property
will restore all the components in the array.