TableState: Remember lowest removed index
Chances are we are going to end up with systems where a lot of entities
are created up front, but most of them are never changed until
destroyed.

Remembering the lowest removed index let's us skipp all the unchanged
entities at the start.

We'll have to reevaluate this at some point and compare it with just
searching for empty slots backwards.