Run exit handlers on removeEntity
Since we're already removing component memory anyway, we can just as
well say that a removed entity loses all its components first.

There seem to be a lot more use-cases where we're tracking combination
of components and a removed entity is just the same as just losing the
components. In fact I can't seem to think of a case where the pedantry
of separating losing all components and removing an entity is a win.

So we run all exit handlers on removeEntity as-if the entity had been
set to mask 0n on all applicable tables first.